fix: Returned false when looking at
When nextIdx is called on node "1.2", it
should fall into this branch and call nextIdx with currentIdx pointing
to "1", and let that "nextIdx" call to determine the final result.
But as it wasn't returned, the call reported -1 no matter what
happens, which creates a corrupted state where it reports operation
failure even when actual m_CurrentIdx is moved forward.

root
|- 1
|  |- 1.1
|  `- 1.2 <-- from here
|
`- 2 <-- to here (expected)